add tsp with time windows with slew time and integration time
added tsp with time windows where the weights are the sum of slew time and integration time.
preop selects first 30 pulsars which will set the earliest
penalties are added

the results are averaged around 55 hours for 30 observations.
it was seen that slew time was increased by adding integration time. This might be because the integration time added to the tsp is the expected integration time. This property changes based on the scintillation and it's like adding weights that are incorrect. Therefore it might dilute the effect of slew time as weights.